Admin pages layout added
