+++ src/Layouts/LayoutSellerM.jsx
@@ -1,0 +1,106 @@
+import * as React from 'react'
+import AdminNavbar from '../components/admin/AdminNavbar'
+import Box from '@mui/material/Box'
+import Button from '@mui/material/Button'
+import SmartTitle from 'components/admin/SmartTitle'
+import { administrador } from '../components/admin/navigationData'
+
+export default function LayoutProdMan({ children }) {
+  // ReactHook para cambiar el estado de la pagina actual
+  const [paginaActual, setPaginaActual] = React.useState(false)
+
+  const ponerEstadoPagina = () => {
+    setPaginaActual((administrador[1].estado = !paginaActual))
+  }
+
+  // ReactHook para cambiar el estado de todas las paginas y de la pagina actual
+  React.useEffect(() => {
+    limpiarPagina()
+    ponerEstadoPagina()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
+  return (
+    <>
+      <AdminNavbar usuario='admin' />
+      <Box sx={{ ...pantallaCompleta }}>
+        {/* Contenedor principal */}
+        <Box sx={{ ...contenedorSuperior }}>
+          {/* Cuadro que contiene a los titlos */}
+          <Box sx={{ ...contenedorTitulos }}>
+            <SmartTitle
+              body={'Gestión de '}
+              bold={'cuentas'}
+              order={''}
+              color={'principal'}
+            />
+          </Box>
+          {/* Cuadro que contiene a los botones */}
+          <Box sx={{ ...contenedorBotones }}>
+            <Button variant='contained' sx={{ ...botonPrimario }}>
+              Nueva cuenta
+            </Button>
+          </Box>
+        </Box>
+        {children}
+      </Box>
+    </>
+  )
+}
+
+const pantallaCompleta = {
+  px: {
+    md: '94px',
+    xs: '18px'
+  }
+}
+
+const contenedorSuperior = {
+  display: 'flex',
+  flexDirection: { xs: 'column', md: 'row' },
+  justifyContent: 'space-between',
+  height: { xs: '55px', md: '143px' },
+  width: '100%',
+  mt: { xs: '10px', md: 0 }
+}
+
+const contenedorTitulos = {
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center'
+}
+
+const contenedorBotones = {
+  position: { xs: 'none', md: 'unset' },
+  bottom: { xs: '21px', md: 0 },
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center',
+  justifyContent: { xs: 'center', sm: 'center', md: 'unset' },
+  mt: { xs: '10px', md: 0 }
+}
+
+const botonPrimario = {
+  backgroundColor: '#FF823B',
+  borderRadius: '10px',
+  width: '270px',
+  height: '39px',
+  textTransform: 'none',
+  fontFamily: 'Noto Sans, sans-serif',
+  fontWeight: '700',
+  fontSize: '1.125rem',
+  boxShadow: 'none',
+  '&:hover': {
+    backgroundColor: '#c65306',
+    boxShadow: 'none'
+  }
+}
+
+// Logica del padre para la barra de navegación
+
+// Función que devuelve el estado de todas las paginas a falso
+const limpiarPagina = () => {
+  for (let i = 0; i < administrador.length; i++) {
+    administrador[i].estado = false
+  }
+}
