--- src/components/admin/SmartButtons.jsx
@@ -1,7 +1,0 @@
-export default function SmartButtons() {
-  return (
-    <>
-      <p></p>
-    </>
-  )
-}
